Hotfix High CPU: Move condition check out of goto loop

+++ script/stop.sh
@@ -1,3 +1,11 @@
 app_pid=`/bin/ps -fu "root" | grep "./bin/app" | grep -v "grep" | awk '{print $2}' | tail -n 1`
-echo Kill ./bin/app with PID= $app_pid
-sudo kill -9 $app_pid
+if [ -n "$app_pid" ]; then
+   echo Kill    app instance with PID = $app_pid
+   sudo kill -9 $app_pid
+fi
+
+app_pid=`/bin/ps -fu "pi" | grep "python " | grep -v "grep" | awk '{print $2}'`
+if [ -n "$app_pid" ]; then
+   echo Kill python instance with PID = $app_pid
+   sudo kill -9 $app_pid
+fi
